Cleaning up warnings with compiler flags -Wall -Wextra -Wpadded Still a few to go in *.cc and threads.c

+++ sources/checkpoint.c
@@ -50,7 +50,7 @@
  *
  */
 /*
-  	#] Explanations : 
+  	#] Explanations :
   	#[ License :
  *
  *   Copyright (C) 1984-2010 J.A.M. Vermaseren
@@ -77,7 +77,7 @@
  *   with FORM.  If not, see <http://www.gnu.org/licenses/>.
  */
 /*
-  	#] License : 
+  	#] License :
   	#[ Includes :
 */
 
@@ -144,7 +144,7 @@ static char *storefile = 0;
 static int done_snapshot = 0;
 
 /*
-  	#] filenames and system commands : 
+  	#] filenames and system commands :
   	#[ CheckRecoveryFile :
 */
 
@@ -248,7 +248,7 @@ int CheckRecoveryFile()
 }
 
 /*
-  	#] CheckRecoveryFile : 
+  	#] CheckRecoveryFile :
   	#[ DeleteRecoveryFile :
 */
 
@@ -288,7 +288,7 @@ void DeleteRecoveryFile()
 }
 
 /*
-  	#] DeleteRecoveryFile : 
+  	#] DeleteRecoveryFile :
   	#[ RecoveryFilename :
 */
 
@@ -301,7 +301,7 @@ char *RecoveryFilename()
 }
 
 /*
-  	#] RecoveryFilename : 
+  	#] RecoveryFilename :
   	#[ InitRecovery :
 */
 
@@ -349,7 +349,7 @@ void InitRecovery()
 }
 
 /*
-  	#] InitRecovery : 
+  	#] InitRecovery :
   	#[ Debugging :
 */
 
@@ -1111,7 +1111,7 @@ static void print_R()
 #endif /* ifdef PRINTDEBUG */
 
 /*
-  	#] Debugging : 
+  	#] Debugging :
   	#[ Cached file operation functions :
 */
 
@@ -1174,7 +1174,7 @@ size_t flush_cache(FILE *fd)
 #endif
 
 /*
-  	#] Cached file operation functions : 
+  	#] Cached file operation functions :
   	#[ Helper Macros :
 */
 
@@ -1207,7 +1207,7 @@ size_t flush_cache(FILE *fd)
 	memcpy(VAR, p, SIZE); p = (unsigned char*)p + SIZE;
 
 #define S_WRITE_B(BUF,LEN) \
-	if ( fwrite_cached(BUF, 1, LEN, fd) != LEN ) return(__LINE__);
+	if ( fwrite_cached(BUF, 1, LEN, fd) != (size_t)(LEN) ) return(__LINE__);
 
 #define S_FLUSH_B \
 	if ( flush_cache(fd) != 1 ) return(__LINE__);
@@ -1223,7 +1223,7 @@ size_t flush_cache(FILE *fd)
 #define S_WRITE_S(STR) \
 	if ( STR ) { \
 		l = strlen((char*)STR) + 1; \
-		if ( fwrite_cached(STR, 1, l, fd) != l ) return(__LINE__); \
+		if ( fwrite_cached(STR, 1, l, fd) != (size_t)l ) return(__LINE__); \
 	}
 
 /* LIST */
@@ -1275,7 +1275,7 @@ time_t announce_time;
 #endif
 
 /*
-  	#] Helper Macros : 
+  	#] Helper Macros :
   	#[ DoRecovery :
 */
 
@@ -1392,7 +1392,7 @@ int DoRecovery(int *moduletype)
 	print_M();
 #endif
 
-	/*#] AM : */ 
+	/*#] AM : */
 	/*#[ AC : */
 
 	/* #[ AC free pointers */
@@ -1913,7 +1913,7 @@ int DoRecovery(int *moduletype)
 	print_C();
 #endif
 
-	/*#] AC : */ 
+	/*#] AC : */
 	/*#[ AP : */
 
 	/* #[ AP free pointers */
@@ -2059,7 +2059,7 @@ int DoRecovery(int *moduletype)
 	print_P();
 #endif
 
-	/*#] AP : */ 
+	/*#] AP : */
 	/*#[ AR : */
 
 	R_SET(ofs,long);
@@ -2234,7 +2234,7 @@ int DoRecovery(int *moduletype)
 	print_R();
 #endif
 
-	/*#] AR : */ 
+	/*#] AR : */
 /*[20oct2009 mt]:*/
 #ifdef PARALLEL
 	/*#[ PF : */
@@ -2254,7 +2254,7 @@ int DoRecovery(int *moduletype)
 	R_SET(PF.exprbufsize, int);
 	R_SET(PF.module, LONG);
 	R_SET(PF.log, int);
-	/*#] PF : */ 
+	/*#] PF : */
 #endif
 /*:[20oct2009 mt]*/
 
@@ -2288,7 +2288,7 @@ int DoRecovery(int *moduletype)
 }
 
 /*
-  	#] DoRecovery : 
+  	#] DoRecovery :
   	#[ DoSnapshot :
 */
 
@@ -2389,7 +2389,7 @@ static int DoSnapshot(int moduletype)
 	S_WRITE_B(&AM.gOldFactArgFlag,sizeof(int));
 	S_WRITE_B(&AM.ggOldFactArgFlag,sizeof(int));
 
-	/*#] AM :*/ 
+	/*#] AM :*/
 	/*#[ AC :*/
 
 	/* we write AC as a whole and then write all additional data step by step.
@@ -2622,7 +2622,7 @@ static int DoSnapshot(int moduletype)
 
 	S_WRITE_S(AC.extrasym);
 
-	/*#] AC :*/ 
+	/*#] AC :*/
 	/*#[ AP :*/
 
 	/* we write AP as a whole and then write all additional data step by step. */
@@ -2708,7 +2708,7 @@ static int DoSnapshot(int moduletype)
 	S_WRITE_B(AP.PreSwitchModes, (AP.NumPreSwitchStrings+1)*(LONG)sizeof(int));
 	S_WRITE_B(AP.PreTypes, (AP.MaxPreTypes+1)*(LONG)sizeof(int));
 
-	/*#] AP :*/ 
+	/*#] AP :*/
 	/*#[ AR :*/
 
 	ANNOUNCE(AR)
@@ -2768,7 +2768,7 @@ static int DoSnapshot(int moduletype)
 	S_WRITE_B(&AR.SortType, sizeof(WORD));
 	S_WRITE_B(&AR.ShortSortCount, sizeof(WORD));
 
-	/*#] AR :*/ 
+	/*#] AR :*/
 
 /*[20oct2009 mt]:*/
 	/*#[ PF :*/
@@ -2780,7 +2780,7 @@ static int DoSnapshot(int moduletype)
 	S_WRITE_B(&PF.module, sizeof(LONG));
 	S_WRITE_B(&PF.log, sizeof(int));
 #endif
-	/*#] PF :*/ 
+	/*#] PF :*/
 /*:[20oct2009 mt]*/
 
 #ifdef WITHPTHREADS
@@ -2850,7 +2850,7 @@ static int DoSnapshot(int moduletype)
 }
 
 /*
-  	#] DoSnapshot : 
+  	#] DoSnapshot :
   	#[ DoCheckpoint :
 */
 
@@ -3001,5 +3001,5 @@ void DoCheckpoint(int moduletype)
 }
 
 /*
-  	#] DoCheckpoint : 
+  	#] DoCheckpoint :
 */
